ParseMonad

Description

Monad for scanning and parsing. The scanner and parser are both monadic, following the design outlined in the Happy documentation on monadic parsers. The parse monad P is built on top of the diagnostics monad D, additionally keeping track of the input and current source code position, and exploiting that the source code position is readily available to avoid having to pass the position as an explicit argument.

Documentation

newtype P a

Constructors

P (Int -> Int -> String -> D a)

unP :: P a -> Int -> Int -> String -> D a

emitInfoP :: String -> P ()

Emits an information message.

emitWngP :: String -> P ()

Emits a warning message.

emitErrP :: String -> P ()

Emits an error message.

failP :: String -> P a

Emits an error message and fails.

getSrcPosP :: P SrcPos

Gets the current source code position.

runP :: P a -> String -> D a

Runs parser (and scanner), yielding a result in the diagnostics monad D.