Scanner

Description

MiniTriangle lexical analyser

Documentation

scanner :: ((Token, SrcPos) -> P a) -> P a

MiniTriangle scanner.

scan :: String -> D [(Token, SrcPos)]

Scans the input and returns the resulting tokens paired with position.

testScanner :: String -> IO ()

Test utility. Scans the input and, if successful, prints the resulting tokens.