Diagnostics

Description

Diagnostic messages and computations (monad). The module provides a datatype for representing diagnostic messages, tagged with source-code position and severity level (error, warning, information). It also provides an abstraction for diagnostic computations, where a computation has the ability to emit diagnostic messages and, possibly, fail. Finally, it provides a utility function for reporting internal compiler errors in a standardized way.

Diagnostic messages

data DMsg

Representation of diagnostic messages

Constructors

DMsg

Fields

dmLvl :: DMLvl

Severity level

dmSrcPos :: SrcPos

Associated source position

dmTxt :: String

Message text

data DMLvl

Severity levels

Constructors

DMLInfo
Information (least severe)

DMLWng
Warning

DMLErr
Error (most severe)

mkInfoMsg :: SrcPos -> String -> DMsg

Constructs information message.

mkWngMsg :: SrcPos -> String -> DMsg

Constructs warning message.

mkErrMsg :: SrcPos -> String -> DMsg

Constructs error message.

dmIsInfo :: DMsg -> Bool

Checks if information message.

dmIsWng :: DMsg -> Bool

Checks if warning message.

dmIsErr :: DMsg -> Bool

Checks if error message.

dmCmpLvlPos :: DMsg -> DMsg -> Ordering

Comparison function for ordering messages first by severity, then by source position.

ppDMsg :: DMsg -> String

Formats a diagnostic message for printing.

Diagnostic computations

data D a

Diagnostic computation. A computation with diagnostic output can succeed or fail, and additionally yields a list of diagnostic messages.

emitInfoD :: SrcPos -> String -> D ()

Emits an information message.

emitWngD :: SrcPos -> String -> D ()

Emits a warning message.

emitErrD :: SrcPos -> String -> D ()

Emits an error message.

failD :: SrcPos -> String -> D a

Emits an error message and fails.

failNoReasonD :: D ()

Fails without giving any specific reason.

failIfErrorsD :: D ()

Fails if there has been errors thus far

stopD :: D a

Forces a stop, e.g. after some user-specified pass.

runD :: D a -> (Maybe a, [DMsg])

Runs a diagnostic computation. Returns:

Result of the computation, if any.

Sorted list of diagnostic messages.

Reporting internal errors

internalError :: String -> String -> String -> a

Signals an internal compiler error. Call with module name, name of function, and error message to report internal errors; i.e., things that should not happen.